+++ install-local.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+code-server --accept-server-license-terms --install-extension $1