+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 All notable changes to the "idle-session-detector" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.2.0] - 2023-10-25
 
-## [Unreleased]
+### Added
+- new 'idleSessionDetector.idleNoticeSeconds' to make the user idle notice interval configurable
+- `install-local.sh` convenience script for installing a locally built `*.vsix`
 
+### Changed
+- `idleNotice`` defaults to 1800 (30 minutes)
+
+## [0.1.0] - 2023-10-23
 - Initial release